aider: Add "start" script to package.json to run "blockchain1.ts" when "npm start" is executed.
# Context:
USER: Modify the package.json """npm start""" script to start `Blockchain1`
ASSISTANT: To modify the `npm start` script to start `Blockchain1`, we will modify the `workspaces/cross-chain_composition/package.json` file.

Here are the changes:

```json
workspaces/cross-chain_composition/package.json
<<<<<<< ORIGINAL
{
  "name": "cross-chain_composition",
  "version": "0.0.1",
  "description": "CPSC 370 assignments",
  "main": "index.ts",
  "scripts": {
    "test": "echo testing"
  },
  "author": "",
  "license": "ISC",
  "dependencies": {
    "crypto": "^1.0.1"
  }
}
=======
{
  "name": "cross-chain_composition",
  "version": "0.0.1",
  "description": "CPSC 370 assignments",
  "main": "index.ts",
  "scripts": {
    "test": "echo testing",
    "start": "ts-node blockchain1.ts"
  },
  "author": "",
  "license": "ISC",
  "dependencies": {
    "crypto": "^1.0.1"
  }
}
>>>>>>> UPDATED
```

In this updated code, the `start` script is added to the `scripts` section of the `package.json` file. The `start` script is set to `ts-node blockchain1.ts`, which will run the `blockchain1.ts` file when `npm start` is executed.