aider: Enforced strict mode in TypeScript by adding a `tsconfig.json` file with the necessary compiler options.
# Context:
USER: do not """use strict""" in the .ts files. Place that in the TypeScipt configuration
ASSISTANT: To enforce strict mode in TypeScript, we need to modify the `tsconfig.json` file. If it doesn't exist, we need to create it.

Here are the changes:

```json
workspaces/cross-chain_composition/tsconfig.json
<<<<<<< ORIGINAL
=======
{
    "compilerOptions": {
        "strict": true,
        "module": "commonjs",
        "target": "es6"
    },
    "include": ["*.ts"]
}
>>>>>>> UPDATED
```

In this updated code, the `strict` option is set to `true` in the `compilerOptions` section of the `tsconfig.json` file. This will enforce strict mode in all TypeScript files. The `module` option is set to `commonjs` and the `target` option is set to `es6` to specify the module system and ECMAScript target version. The `include` option is set to `["*.ts"]` to include all TypeScript files in the project.

Please note that you need to remove `"use strict";` from the top of each `.ts` file since it's not needed when `strict` is enabled in `tsconfig.json`.