aider: Removed direct calls to createBlock method of both blockchains and removed logs of the blockchain chains and shared state.
# Context:
USER: Use only the exports from the `workspaces/cross-chain_composition/blockchain1.ts` and `workspaces/cross-chain_composition/blockchain2.ts` files in the `workspaces/cross-chain_composition/interblockchain.ts`. Allow these two files to independently execute their own processes in the block production and issue updates to the `workspaces/cross-chain_composition/interblockchain.ts` process.
ASSISTANT: To achieve this, we need to modify the `workspaces/cross-chain_composition/interblockchain.ts` file. We will remove the direct calls to `createBlock` method of both blockchains and instead, we will rely on the blockchains to independently execute their own processes in block production. We will also remove the logs of the blockchain chains and shared state, as they will be updated independently by the blockchains.

Here is the change:

```typescript
workspaces/cross-chain_composition/interblockchain.ts
<<<<<<< ORIGINAL
        blockchain1.createBlock({ info: 'New block in Blockchain1' });
        blockchain2.createBlock({ info: 'New block in Blockchain2' });

        console.log('Blockchain1:', blockchain1.chain);
        console.log('Blockchain2:', blockchain2.chain);
        console.log('Shared State:', sharedState);
=======
        // Removed direct calls to createBlock method of both blockchains
        // Removed logs of the blockchain chains and shared state
>>>>>>> UPDATED
```

Please note that this change assumes that the `blockchain1.ts` and `blockchain2.ts` files have their own logic for block production and updating the shared state in the `interblockchain.ts` file. If this is not the case, we would need to modify those files as well.