aider: Calculate the length of the chains from the log files in the `InterBlockchain` function.
# Context:
USER: have `InterBlockchain` obtain the chain length from the log files and use that file solely for the interchain communication
ASSISTANT: To achieve this, we need to modify the `InterBlockchain` function in the `interblockchain.ts` file. We will read the log files and calculate the length of the chains from the number of blocks in the log files.

```typescript
workspaces/cross-chain_composition/interblockchain.ts
<<<<<<< ORIGINAL
    const InterBlockchain = () => {
        console.log('InterBlockchain function called');
        const sharedState: any = {};

        let lastBlockTime: number | null = null;

    const fs = require('fs');
    const logFile1 = 'blockchain1.log';
    const logFile2 = 'blockchain2.log';
=======
    const InterBlockchain = () => {
        console.log('InterBlockchain function called');
        const sharedState: any = {};

        let lastBlockTime: number | null = null;

    const fs = require('fs');
    const logFile1 = 'blockchain1.log';
    const logFile2 = 'blockchain2.log';

    let chain1Length = fs.readFileSync(logFile1, 'utf-8').split('\n').filter(Boolean).length;
    let chain2Length = fs.readFileSync(logFile2, 'utf-8').split('\n').filter(Boolean).length;
>>>>>>> UPDATED
```

In the updated code, we read the log files and split the content by newline characters to get an array of blocks. We then filter out any empty strings (which can occur if there's a newline at the end of the file), and get the length of the resulting array to find the number of blocks in each chain.