aider: Replace the React logo with the Chapman class one.
# Aider chat conversation:

USER: replace the React logo with the Chapman class one

ASSISTANT: In order to assist you with this, I need the React project files where the logo is being used. Typically, this would be in a file like `src/App.js` or `src/App.tsx` and the logo file would be `src/logo.svg`.

Please add these files to the chat so I can propose the necessary changes. Also, please provide the Chapman class logo file.

USER: `workspaces/mySwap/src/Header.tsx`

ASSISTANT: To replace the React logo with the Chapman class logo in the `Header.tsx` file, we need to change the source of the image in the `img` tag.

Here is the proposed change:

```tsx
workspaces/mySwap/src/Header.tsx
<<<<<<< HEAD
      <img src={logo} alt="Chapman class logo" />
=======
      <img src="./chapman_class_logo.svg" alt="Chapman class logo" />
>>>>>>> updated
```

This change will replace the current logo with the Chapman class logo. Please ensure that the `chapman_class_logo.svg` file is located in the same directory as the `Header.tsx` file. If it's in a different directory, adjust the path accordingly.