aider: Added React, ReactDOM, and create-react-app packages to the project and added a script to start the React app in the `package.json` file.
# Aider chat conversation:

ASSISTANT: Sure, to set up a basic interface for this project using ReactJS, we need to add React and its related packages to the project. We will also need to add a script to start the React app in the `package.json` file.

Here are the steps:

1. Add React, ReactDOM, and create-react-app to the project. This can be done by adding these packages to the dependencies in the `package.json` file.

2. Add a script to start the React app. This can be done by adding a "start" script in the scripts section of the `package.json` file.

Here is the *edit block* for the `package.json` file:

```json
workspaces/codingProject2/package.json
<<<<<<< HEAD
{
  "name": "codingproject2",
  "version": "1.0.0",
  "description": "",
  "main": "index.js",
  "scripts": {
    "test": "echo \"Error: no test specified\" && exit 1"
  },
  "author": "",
  "license": "ISC"
}
=======
{
  "name": "codingproject2",
  "version": "1.0.0",
  "description": "",
  "main": "index.js",
  "scripts": {
    "start": "react-scripts start",
    "test": "echo \"Error: no test specified\" && exit 1"
  },
  "dependencies": {
    "react": "^17.0.2",
    "react-dom": "^17.0.2",
    "react-scripts": "4.0.3"
  },
  "author": "",
  "license": "ISC"
}
>>>>>>> updated
```

After this, we need to create the entry point for the React app, which is usually the `index.js` file. However, I don't see the `index.js` file in the list of *read-write* files you provided. Could you please add it to the chat?