aider: Import `Lib` from `strict_types` in `schema.rs` to resolve the issue of `Lib` not being found in the scope.
# Aider chat conversation:

USER: In "static LIB: Result<Lib, Box<dyn Error>> = LibBuilder::new(libname!(LIB_NAME_RGB_CONTRACT))
    .process::<Nominal>()" it says that the variable Lib is not found in this scope.

ASSISTANT: To resolve the issue of the `Lib` type not being found in the scope, we need to ensure that it is correctly imported from the appropriate crate or module. If `Lib` is supposed to be part of the `strict_types` crate or another crate that is already included in the dependencies, we should import it at the top of the `schema.rs` file.

However, since I don't have information about where `Lib` is defined, I'll assume for now that it's part of the `strict_types` crate, which seems to be imported in the file. If `Lib` is not part of `strict_types`, you will need to provide the correct crate or module name.

Let's add the import statement for `Lib` to the `schema.rs` file:

```rust
workspaces/rgb20-token/src/schema.rs
<<<<<<< SEARCH
use strict_types::LibBuilder;
=======
use strict_types::{Lib, LibBuilder};
>>>>>>> REPLACE
```

If `Lib` is not part of `strict_types` and is instead part of a different crate that is not yet included in the dependencies, you will need to add that crate to `Cargo.toml` and then import `Lib` from that crate. If that's the case, please provide the correct crate name, and I can guide you through the process of adding it to `Cargo.toml` and importing `Lib`.